backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
replicate==0.25.1
python-multipart==0.0.6
Pillow==10.1.0
python-dotenv==1.0.0

